add white list

+++ user.py
@@ -41,7 +41,7 @@ ai_client = genai.Client(api_key=api_key)
 
 
 def not_is_white(id_user):
-    with ('config.json', 'r') as file:
+    with open('config.json', 'r') as file:
         bd = json.load(file)
         if id_user in bd['white']:
             return False
@@ -78,6 +78,8 @@ async def host(client:Client, message:Message):
 
 @client.on_message(filters.command(commands=['audio', 'поясни','аудио'],prefixes=['!','.']))
 async def wav(client:Client, message:Message):
+    if not_is_white(message.from_user.id):
+        return #перевірка чи немає користувача в білому списку, якшо немає видає True
     print(1)
     print(message.reply_to_message.media)
     print(message.reply_to_message,message.reply_to_message.media )
@@ -108,21 +110,33 @@ async def wav(client:Client, message:Message):
 
 @client.on_message(filters.command(commands=['list', 'white','білий'],prefixes=['!','.']))
 async def white_list(client:Client, message:Message):
+    print(not_is_white(message.from_user.id))
+    if not_is_white(message.from_user.id):
+        return #перевірка чи немає користувача в білому списку, якшо немає видає True
+    text = message.text.split()
+    if len(text) == 1:
+            with open('config.json', 'r') as file:
+                bd = json.load(file)
+                await message.reply(bd['white'])
     if message.reply_to_message != None:
-        text = message.text.split()
+        
         mode = text[1]
+        print(mode)
         if str(mode) == "add":
-            add_white(message.reply_to_message.id)
+            add_white(message.reply_to_message.from_user.id)
             await message.react("👍")
             
-        if str(mode) == "del":
-            del_white(message.reply_to_message.id)
+        elif str(mode) == "del":
+            print(1)
+            del_white(message.reply_to_message.from_user.id)
             await message.react("👍")
             
 
 
 @client.on_message(filters.command(commands=['setmodel', 'сетмодел','сетмодель','сетгеміні','модель','model','setgemini'],prefixes=['!','.']))
 async def setmodel(client:Client, message:Message):
+        if not_is_white(message.from_user.id):
+            return #перевірка чи немає користувача в білому списку, якшо немає видає True
         text = message.text.split()
         if len(text) == 1:
             with open('config.json', 'r') as file:
@@ -161,6 +175,8 @@ async def unban(client:Client,message:Message):
 
 @client.on_message(filters.command(commands=['ban', 'block','блок','бан'],prefixes=['!','.']))
 async def ban(client:Client,message:Message):
+            if not_is_white(message.from_user.id):
+                return #перевірка чи немає користувача в білому списку, якшо немає видає True
             id_user = message.reply_to_message.from_user.id
             await client.block_user(id_user)
             await message.react("👍")
@@ -195,7 +211,9 @@ async def url(client: Client, message: Message):
                 await message.react("🙈")
                 await message.reply(e)
 @client.on_message(filters.command(commands=['name', 'nick','нік'],prefixes=['!','.']))
-async def name(client: Client, message: Message): 
+async def name(client: Client, message: Message):
+        if not_is_white(message.from_user.id):
+            return #перевірка чи немає користувача в білому списку, якшо немає видає True
         user = message.text.split()
         user = user[1:]
         user = ' '.join(user)
@@ -213,32 +231,57 @@ async def name(client: Client, message: Message):
 @client.on_message(filters.command(commands=['муз', 'mus','music'],prefixes=['!','.']))
 async def mus(client: Client, message: Message): 
         url = message.text.split()
-        try:
-            if len(url) >= 2:
+        if len(url) >= 2:
             
-                url = url[1:]
-                await message.react("👀")
-                await client.send_message('LyBot', text=f"Знайди пісню:{" ".join(url)}")
-                await asyncio.sleep(5)
-                async for i in client.get_chat_history(chat_id="LyBot", limit=1):
-                    print(i)
-                    if i.audio != None or i.text != None:
-                        if i.text == None:
-                            await client.forward_messages(message.chat.id,i.chat.id, i.id)
-                        else:
-                            if "Для завантаження" in i.text :
-                                await message.reply("Пісню не можливо завантажити")
-                                return
-                            text = "Пісня не знайдена може ви мали на увазі:"
-                            for num in range(0,5):
-                                text += f"\n{i.reply_markup.inline_keyboard[num][0].text}"
+            url = url[1:]
+            await message.react("👀")
 
-                            await message.reply(text)
-                        await message.react('')
+            print("for")
+            succes = False
+            send = 'LyBot'
+            kb = False
+            while succes != True:
+                try:
+               
+                    if send == "LyBot" and kb == False:
+                        await client.send_message('LyBot', text=f"Знайди пісню:{" ".join(url)}")
+                    elif kb == False :
+                        await message.reply('Не вдалося завантажити з лубота. Вантажу з іншого джерела')
+                        await client.send_message('vkmusic_bot', text=f"{" ".join(url)}")
+                    await asyncio.sleep(5)
+                    async for i in client.get_chat_history(chat_id=send, limit=1):
+                        if i.audio != None or i.text != None:
+                            print(i,i.audio)
+                            if i.text == None:
+                                print(3)
+                                await client.forward_messages(message.chat.id,i.chat.id, i.id)
+                                succes = True
+                            elif "Для завантаження" in i.text :
+                                    send = "vkmusic_bot"              
+                                    
+                            else:
+                                
+                                print('1', i.audio == None)
+                                #натискаємо на першу кнопку якщо дають декілька варінтів
+                                callback_but1 = i.reply_markup.inline_keyboard[0][0].callback_data #["videold"]
+                                print(type(callback_but1))
+                                request = await client.request_callback_answer(
+                            chat_id=i.chat.id,
+                            message_id=i.id,
+                            callback_data=callback_but1)
+                                kb = True
+                                if send == "LyBot":
+                                    pass
+                                else:
+                                    send = "vkmusic_bot"
+                                await asyncio.sleep(2)
+                                                
+                except Exception as e:
+                    
+                    send = "vkmusic_bot"
+        await message.react("👍")
                 
-        except Exception as e:
-                await message.reply(e)
-                print(e)
+                                
 
                 
 @client.on_message(filters.command(commands=['карта', 'map','тривога'],prefixes=['!','.']))
@@ -264,6 +307,8 @@ async def text(client:Client,message):
 
 @client.on_message(filters.command('ем',prefixes=['!', '.']))
 async def em(client:Client,message):
+    if not_is_white(message.from_user.id):
+        return #перевірка чи немає користувача в білому списку, якшо немає видає True
     id_chat = message.chat.id 
     await client.delete_messages(id_chat,message.id)
     emojis = [
@@ -291,6 +336,8 @@ async def em(client:Client,message):
 
 @client.on_message(filters.command('сер',prefixes=['!', '.']))
 async def em(client:Client, message:Message):
+        if not_is_white(message.from_user.id):
+            return #перевірка чи немає користувача в білому списку, якшо немає видає True
         id_chat = message.chat.id 
         heart_emojis = ["❤️", "🧡",  "💛",  "💚", "💙",  "💜",  "🤎",  "🖤", "🤍",  "💖",  "💗", "💘", "❤️‍🔥",  "❤️‍🩹"]
         await client.delete_messages(id_chat,message.id)
@@ -305,6 +352,8 @@ async def em(client:Client, message:Message):
 
 @client.on_message(filters.command(commands=['ai','.', 'ші',],prefixes=['!', '.']))
 async def ai(client:Client,message:Message):
+    if not_is_white(message.from_user.id):
+        return #перевірка чи немає користувача в білому списку, якшо немає видає True
     try:
             print(message)
             #розбиваємо текст на елменти списку
@@ -341,6 +390,35 @@ async def ai(client:Client,message:Message):
     except Exception as e:
         await message.react("🙈")
         await message.reply(text=e)
+
+@client.on_message(filters.command(commands=['del', 'дел','вид'],prefixes=['!','.']))
+async def delete(client:Client,message:Message):
+    if message.chat.type !='private':
+        user = await client.get_users('aHrel_cMePTI')
+        print(user)
+        members = client.get_chat_members(message.chat.id)
+        print(members)
+        deleted = 0
+        premium = 0
+        bot = 0
+        normal = 0
+        async for i in members:
+            print(i)
+            if i.user.is_deleted == True:
+                deleted += 1
+            elif i.user.is_premium == True:
+                 premium += 1
+            elif i.user.is_bot == True:
+                 bot += 1
+            else:
+                 normal += 1
+        text = f"""У цьому чаті всього {sum([deleted,premium,bot,normal])} користувачів
+з них:
+{deleted} - видалені акаунти
+{bot} - боти
+{premium} - мають преміум
+"""
+        await message.reply(text)
